+++ cmd/oh/segment/modeToggle/modeToggle.go
@@ -36,7 +36,7 @@ func (self state) Render(segment.Context) string {
 		self.letter(caps.Write, grantedCaps.Has(caps.Write), style.Write, isPrefixPending) +
 		gap +
 		self.letter(caps.Network, grantedCaps.Has(caps.Network), style.Network, isPrefixPending) +
-		self.letter(caps.Git, grantedCaps.Has(caps.Git), style.History, isPrefixPending) +
+		self.letter(caps.Git, grantedCaps.Has(caps.Git), style.Git, isPrefixPending) +
 		self.letter(caps.Lookup, grantedCaps.Has(caps.Lookup), style.Lookup, isPrefixPending)
 }